% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBORoptim.R
\name{PBORoptim}
\alias{PBORoptim}
\title{Prior Based Optimization Routine}
\usage{
PBORoptim(
  x0,
  func,
  lb,
  ub,
  center_vec,
  sigma_vec,
  max_time = 6000,
  rho = 2,
  phi = 10^(-20),
  no_runs = 10,
  max_iter = 10000,
  s_init = 1,
  tol_fun = 10^(-20),
  tol_fun_2 = 10^(-20)
)
}
\arguments{
\item{x0}{Initial guess (vector) of solution.}

\item{func}{Function needs to be minimized.}

\item{lb}{Lower bound vector of parameter domain.}

\item{ub}{Upper bound vector of domain.}

\item{center_vec}{Mean vector of prior normal distribution. It can be set at the
neighborhood where the true parameter value is expected. Default value is 0 vector.}

\item{sigma_vec}{The standard deviation vector of the prior normal distribution.}

\item{max_time}{Maximum time alloted for optimization.}

\item{rho}{Step Decay Rate with default value 2}

\item{phi}{Step size threshold, i.e., lower Bound Of Global Step Size. Default value is \eqn{10^{-20}}}

\item{no_runs}{Maximum number of runs to be executed, default value is 1000.}

\item{max_iter}{Max Number Of Iterations in each Run. Default Value is 10000.}

\item{s_init}{Initial Global Step Size. Default Value is 1.}

\item{tol_fun}{Termination Tolerance on the function value. Default Value is \eqn{10^{-20}}}

\item{tol_fun_2}{Termination Tolerance on the difference of solutions in two consecutive
runs. Default Value is \eqn{10^{-20}}.}
}
\value{
The point where func is minimized.
}
\description{
\code{PBOR} function minimizes Black-box function of parameters where each parameter belongs
to any one of this kind of intervals : -Inf to Inf, -Inf to F, F to Inf, F to G; where
F and G are finite known numbers.
}
\examples{
func <- function(x){return(100*(x[2]-x[1]^2)^2 + (x[1]-1)^2)}
PBORoptim(x0 = c(0,0), func, lb = c(-100,-Inf), ub = c(10, Inf), center_vec = c(0,0),
sigma_vec = c(10^2, 10^2))
}
\references{
Das, P. and Ghosh, A. (2021). \emph{PBOR paper}. Future Journal, 2, 63-74.
}
\author{
Priyam Das (maintainer) (\email{priyam_das@hms.harvard.edu}), Abhik Ghosh.
}
